convert Ati FT. Still raw gauge data.
needs calibration slot setting and scaling.

+++ src/MojoClient.cpp
@@ -188,12 +188,12 @@ void MojoClient::rosPublish(){
     }
 
     if (USE_AXIA_FT_SENSOR) {
-        msg_ati_ft.wrench.force.x = robotState().getDigitalIOValue("AtiAxiaFtSensor.Fx");
-        msg_ati_ft.wrench.force.y = robotState().getDigitalIOValue("AtiAxiaFtSensor.Fy");
-        msg_ati_ft.wrench.force.z = robotState().getDigitalIOValue("AtiAxiaFtSensor.Fz");
-        msg_ati_ft.wrench.torque.x = robotState().getDigitalIOValue("AtiAxiaFtSensor.Tx");
-        msg_ati_ft.wrench.torque.y = robotState().getDigitalIOValue("AtiAxiaFtSensor.Ty");
-        msg_ati_ft.wrench.torque.z = robotState().getDigitalIOValue("AtiAxiaFtSensor.Tz");
+        msg_ati_ft.wrench.force.x = (float) (int) robotState().getDigitalIOValue("AtiAxiaFtSensor.Fx");
+        msg_ati_ft.wrench.force.y = (float) (int) robotState().getDigitalIOValue("AtiAxiaFtSensor.Fy");
+        msg_ati_ft.wrench.force.z = (float) (int) robotState().getDigitalIOValue("AtiAxiaFtSensor.Fz");
+        msg_ati_ft.wrench.torque.x = (float) (int) robotState().getDigitalIOValue("AtiAxiaFtSensor.Tx");
+        msg_ati_ft.wrench.torque.y = (float) (int) robotState().getDigitalIOValue("AtiAxiaFtSensor.Ty");
+        msg_ati_ft.wrench.torque.z = (float) (int) robotState().getDigitalIOValue("AtiAxiaFtSensor.Tz");
     }
 
     if (USE_MEDIA_FLANGE) {
